Implement better error handling for project-builder service.

+++ services/project_builder_service.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import subprocess
 import tempfile
 import time
@@ -7,23 +8,29 @@ import cxxd.service
 class ProjectBuilder(cxxd.service.Service):
     def __init__(self, service_plugin):
         cxxd.service.Service.__init__(self, service_plugin)
-        self.build_cmd_dir = ""
-        self.build_cmd_output_file = ""
+        self.build_cmd_dir = None
+        self.build_cmd_output_file = None
 
     def startup_callback(self, args):
-        self.build_cmd_dir = args[0]
-        self.build_cmd_output_file = tempfile.NamedTemporaryFile(suffix='_project_build_output', delete=True)
-        logging.info("Args = {0}, build_cmd_output_file = {1}.".format(args, self.build_cmd_output_file.name))
+        build_cmd_dir = args[0]
+        if os.path.isdir(build_cmd_dir):
+            self.build_cmd_dir = build_cmd_dir
+            self.build_cmd_output_file = tempfile.NamedTemporaryFile(suffix='_project_build_output')
+            logging.info("Build command will be executed from \'{0}\' directory. Output will be recorded into \'{1}\'.".format(build_cmd_dir, self.build_cmd_output_file.name))
+        else:
+            logging.error('Provided directory, \'{0}\', is not valid!'.format(build_cmd_dir))
 
     def shutdown_callback(self, args):
         pass
 
     def __call__(self, args):
-        start = time.clock()
         build_cmd = args[0]
-        self.build_cmd_output_file.truncate()
-        cmd = "cd " + self.build_cmd_dir + " && " + build_cmd
-        ret = subprocess.call(cmd, shell=True, stdout=self.build_cmd_output_file, stderr=self.build_cmd_output_file)
-        end = time.clock()
-        logging.info("Cmd '{0}' took {1}".format(cmd, end-start))
-        return self.build_cmd_output_file.name, end-start
+        if self.build_cmd_dir and self.build_cmd_output_file:
+            start = time.clock()
+            self.build_cmd_output_file.truncate()
+            cmd = "cd " + self.build_cmd_dir + " && " + build_cmd
+            ret = subprocess.call(cmd, shell=True, stdout=self.build_cmd_output_file, stderr=self.build_cmd_output_file)
+            end = time.clock()
+            logging.info("Cmd '{0}' took {1}".format(cmd, end-start))
+            return ret == 0, [self.build_cmd_output_file.name, end-start]
+        return False, None
